requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
identify==2.6.0
importlib-metadata==4.13.0
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.8.0
psycopg==3.2.1
PyJWT==2.9.0
PyYAML==6.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.26.3
zipp==3.19.2
